+++ include/Options/Abstract_Option.php
@@ -6,6 +6,7 @@
 namespace WP_Syntex\Polylang\Options;
 
 use WP_Error;
+use WP_Term;
 use WP_Syntex\Polylang\Options\Options;
 
 defined( 'ABSPATH' ) || exit;
@@ -225,4 +226,58 @@ abstract class Abstract_Option {
 	 * @return string
 	 */
 	abstract protected function get_description(): string;
+
+	/**
+	 * Returns a list of language terms.
+	 *
+	 * @since 3.7
+	 *
+	 * @param string $fields Optional. Term fields to query for (see `WP_Term_Query` for possible values). Default is `'all'`.
+	 * @return array
+	 *
+	 * @phpstan-param 'all'|'all_with_object_id'|'names'|'slugs'|'id=>name'|'id=>slug'|'ids'|'tt_ids' $fields
+	 * @phpstan-return (
+	 *     $fields is 'all'|'all_with_object_id' ? list<WP_Term> : (
+	 *         $fields is 'names'|'slugs'|'id=>name'|'id=>slug' ? array<int, string> : array<int, int>
+	 *     )
+	 * )
+	 */
+	protected function get_language_terms( string $fields = 'all' ): array {
+		$language_terms = get_terms(
+			array(
+				'taxonomy'   => 'language',
+				'hide_empty' => false,
+				'fields'     => $fields,
+			)
+		);
+		return is_array( $language_terms ) ? $language_terms : array();
+	}
+
+	/**
+	 * Adds a non-blocking error warning about unknown language slugs.
+	 *
+	 * @since 3.7
+	 *
+	 * @param array $language_slugs List of language slugs.
+	 * @return void
+	 */
+	protected function add_unknown_languages_warning( array $language_slugs ): void {
+		$this->errors->add(
+			sprintf( 'pll_unknown_%s_languages', static::key() ),
+			sprintf(
+				/* translators: %s is a list of language slugs. */
+				_n( 'The language %s is unknown and has been discarded.', 'The languages %s are unknown and have been discarded.', count( $language_slugs ), 'polylang' ),
+				wp_sprintf_l(
+					'%l',
+					array_map(
+						function ( $slug ) {
+							return "<code>{$slug}</code>";
+						},
+						$language_slugs
+					)
+				)
+			),
+			'warning'
+		);
+	}
 }
