+++ include/Options/Business/Domains.php
@@ -107,13 +107,7 @@ class Domains extends Abstract_Option {
 		/** @phpstan-var array<non-falsy-string, string> $value */
 		$all_values     = array(); // Previous and new values.
 		$missing_langs  = array(); // Lang names corresponding to the empty values.
-		$language_terms = get_terms(
-			array(
-				'taxonomy'   => 'language',
-				'hide_empty' => false,
-			)
-		);
-		$language_terms = is_array( $language_terms ) ? $language_terms : array();
+		$language_terms = $this->get_language_terms();
 
 		// Detect empty values, fill missing keys with previous values.
 		foreach ( $language_terms as $lang ) {
@@ -135,25 +129,7 @@ class Domains extends Abstract_Option {
 		// Detect invalid language slugs.
 		if ( ! empty( $value ) ) {
 			// Non-blocking error.
-			$value = array_keys( $value );
-
-			$this->errors->add(
-				'pll_unknown_domain_languages',
-				sprintf(
-					/* translators: %s is a list of language slugs. */
-					_n( 'The language %s is unknown and has been discarded.', 'The languages %s are unknown and have been discarded.', count( $value ), 'polylang' ),
-					wp_sprintf_l(
-						'%l',
-						array_map(
-							function ( $slug ) {
-								return "<code>{$slug}</code>";
-							},
-							$value
-						)
-					)
-				),
-				'warning'
-			);
+			$this->add_unknown_languages_warning( array_keys( $value ) );
 		}
 
 		if ( 3 === $options->get( 'force_lang' ) && ! empty( $missing_langs ) ) {
